Inline protractor.tsx into stateful-protractor.tsx
