+++ packages/perseus/src/widgets/interactive-graphs/protractor/stateful-protractor.tsx
@@ -1,12 +1,27 @@
-import {useTransformContext, vec} from "mafs";
+import {vec} from "mafs";
 import * as React from "react";
 import {RefObject, useRef, useState} from "react";
 
 import {useDraggable} from "../graphs/use-draggable";
-import {centerToRotationHandle, Protractor} from "./protractor";
 import useGraphConfig from "../reducer/use-graph-config";
-import {bound} from "../utils";
+import {bound, TARGET_SIZE} from "../utils";
 import {useDrag} from "@use-gesture/react";
+import {useTransformVectorsToPixels} from "../graphs/use-transform";
+import {calculateAngleInDegrees} from "../graphs/utils";
+import {pathBuilder} from "../../../util/svg";
+
+import "./protractor.css";
+
+const protractorImage =
+    "https://ka-perseus-graphie.s3.amazonaws.com/e9d032f2ab8b95979f674fbfa67056442ba1ff6a.png";
+
+// The vector from the center of the protractor to the top left corner of the
+// protractor image, in pixels. Used for positioning.
+const centerToTopLeft: vec.Vector2 = [-180, -170];
+
+// The vector from the center of the protractor to the center of the rotation
+// handle.
+export const centerToRotationHandle: vec.Vector2 = [-176, -15];
 
 export function StatefulProtractor() {
     const {range, snapStep} = useGraphConfig();
@@ -37,13 +52,53 @@ export function StatefulProtractor() {
         constrain: constrainToCircle,
     });
 
+    const [centerPx] = useTransformVectorsToPixels(center);
+    const topLeftPx = vec.add(centerPx, centerToTopLeft);
+    const angle = calculateAngleInDegrees(rotationHandleOffset) - calculateAngleInDegrees(centerToRotationHandle);
+
     return (
-        <g ref={draggableRef}>
-            <Protractor
-                rotationHandleRef={rotationHandleRef}
-                rotationHandleOffset={rotationHandleOffset}
-                center={center}
+        <g
+            ref={draggableRef}
+            transform={`translate(${topLeftPx[0]}, ${topLeftPx[1]}), rotate(${angle})`}
+            style={{transformOrigin: "180px 170px"}}
+        >
+            <image href={protractorImage} />
+            <g transform={`translate(5, 170)`} ref={rotationHandleRef}>
+                <RotationArrow />
+            </g>
+        </g>
+    );
+}
+
+function RotationArrow() {
+    const radius = 175;
+    const angleDeg = 10;
+    const angleRad = degreesToRadians(angleDeg);
+    const endX = radius * (1 - Math.cos(angleRad));
+    const endY = radius * -Math.sin(angleRad);
+    const rotationArrow = pathBuilder()
+        .move(0, 0)
+        .circularArc(radius, endX, endY, {sweep: true})
+        .build();
+
+    const arrowhead = pathBuilder().move(-8, 0).line(0, 10).line(8, 0).build();
+
+    const targetRadius = TARGET_SIZE / 2;
+    return (
+        <g className="protractor-rotation-handle">
+            <path className="protractor-rotation-handle-arrow-arc"
+                  d={rotationArrow}
             />
+            <path className="protractor-rotation-handle-arrowhead"
+                  d={arrowhead}
+            />
+            <path className="protractor-rotation-handle-arrowhead"
+                  d={arrowhead}
+                  transform={`translate(${endX}, ${endY}), rotate(${180 + angleDeg})`}
+            />
+            {/* this invisible ellipse ensures that the click target for the
+              * handle is at least 48x48 pixels */}
+            <ellipse cx="0px" cy="-15px" rx={targetRadius} ry={targetRadius} fill="none"></ellipse>
         </g>
     );
 }
@@ -109,4 +164,8 @@ function useDraggablePx(args: {
 // - if `fraction` is 1, `lerp` returns `b`.
 function lerp(a: number, b: number, fraction: number): number {
     return (b - a) * fraction + a;
+}
+
+function degreesToRadians(degrees) {
+    return (degrees / 180) * Math.PI;
 }
